+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - `AbstractMoney::compareTo()`, `isEqualTo()`, `isLessThan()`, `isLessThanOrEqualTo()`, `isGreaterThan()`, `isGreaterThanOrEqualTo()`
   - `ConfigurableProvider::setExchangeRate()`
 - **Calling `Currency::of()` with a numeric code is no longer supported**, use `Currency::ofNumericCode()` instead (#104)
+- `Currency::is()` has been removed, use `Currency::isEqualTo()` instead
 
 ## [0.11.1](https://github.com/brick/money/releases/tag/0.11.1) - 2026-02-12